refractor and mypy stuff
